edit fragments
