+++ app/src/main/java/com/genrikhsalexandr/androidintesive/FragmentA.kt
@@ -1,6 +1,5 @@
 package com.genrikhsalexandr.androidintesive
 
-import FragmentB
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -8,7 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import com.genrikhsalexandr.androidintesive.databinding.FragmentABinding
 
-class AFragment : Fragment() {
+class FragmentA : Fragment() {
 
     private var _binding: FragmentABinding? = null
     private val binding: FragmentABinding get() = _binding!!
@@ -19,13 +18,16 @@ class AFragment : Fragment() {
     ): View {
         _binding = FragmentABinding.inflate(inflater, container, false)
         binding.btFragmentB.setOnClickListener {
-            val fragmentB = FragmentB()
-            val transaction = requireActivity().supportFragmentManager.beginTransaction()
-            transaction.replace(R.id.containerFragment, fragmentB)
-            transaction.addToBackStack(null)
-            transaction.commit()
+            fragmentB()
         }
         return binding.root
+    }
+
+    private fun fragmentB() {
+        requireActivity().supportFragmentManager.beginTransaction()
+            .replace(R.id.containerFragment, FragmentB.newInstance())
+            .addToBackStack(null)
+            .commit()
     }
 
     override fun onDestroyView() {
