+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/containerFragment"
-    android:name="com.genrikhsalexandr.androidintesive.AFragment"
+    android:name="com.genrikhsalexandr.androidintesive.FragmentA"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity"
