+++ app/src/main/res/layout/fragment_a.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".AFragment">
+    tools:context=".FragmentA">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/nameFragment"
